Move Sucess message logic into ContactForm
