+++ pages/kontakt.js
@@ -1,19 +1,9 @@
 import Layout from '../components/Layout'
 import kontakt from '../assets/api/api-kontakt'
 import HeadlineSvgInfo from '../components/HeadlineSvgInfo/HeadlineSvgInfo'
-import Logo from '../components/Logo/Logo'
 import ContactForm from '../components/ContactForm/ContactForm'
-import styled from 'styled-components'
-import MessageSendSVG from '../assets/svgs/kontakt/emailSend.svg'
-import { useState } from 'react'
 
 export default function Kontakt() {
-  const [isSend, setIsSend] = useState(false)
-
-  const onSubmit = () => {
-    setIsSend(true)
-  }
-
   return (
     <Layout>
       {kontakt.map((element, index) => (
@@ -26,34 +16,10 @@ export default function Kontakt() {
         </HeadlineSvgInfo>
       ))}
 
-      {isSend ? (
-        <>
-          <MessageSendSVG />
-          <SucessMessage>
-            Wir haben ihre Nachricht erhalten und melden uns umgehend bei Ihnen!
-          </SucessMessage>
-        </>
-      ) : (
-        <>
-          <Headline>
-            Zu welchem der beiden Themen wollen Sie uns schreiben?
-          </Headline>
-          <ContactForm onSubmit={onSubmit} />
-        </>
-      )}
+      <ContactForm />
     </Layout>
   )
 }
-
-const Headline = styled.p`
-  font-weight: bold;
-  text-align: center;
-  font-size: 1.1em;
-`
-
-const SucessMessage = styled.h4`
-  text-align: center;
-`
 
 //  {/* <form onSubmit={handleSubmit(onSubmit)}>
 //         {/* register your input into the hook by invoking the "register" function */}
